extract _pg to separate module
